#1114 コンポーネント分割
