+++ src/components/Sing/ScoreSequencer.vue
@@ -1,74 +1,40 @@
 <template>
   <div class="score-sequencer" id="score-sequencer">
-    <!-- 鍵盤表示 -->
-    <svg
-      width="48"
-      v-bind:height="`${BASE_Y_SIZE * zoomY * 128}`"
-      xmlns="http://www.w3.org/2000/svg"
-      class="sequencer-keys"
-    >
-      <g v-for="(y, index) in gridY" :key="index">
-        <rect
-          x="0"
-          v-bind:y="`${BASE_Y_SIZE * zoomY * index}`"
-          v-bind:width="`${y.color === 'black' ? 30 : 48}`"
-          v-bind:height="`${BASE_Y_SIZE * zoomY}`"
-          v-bind:class="`sequencer-keys-item-${y.color}`"
-          v-bind:title="y.name"
-        />
-        <line
-          x1="0"
-          v-bind:y1="`${(index + 1) * BASE_Y_SIZE * zoomY}`"
-          x2="48"
-          v-bind:y2="`${(index + 1) * BASE_Y_SIZE * zoomY}`"
-          stroke-width="1"
-          v-bind:class="`sequencer-keys-item-separator ${
-            y.pitch === 'C' && 'sequencer-keys-item-separator-octave'
-          } ${y.pitch === 'F' && 'sequencer-keys-item-separator-f'}`"
-          v-if="y.pitch === 'C' || y.pitch === 'F'"
-        />
-        <text
-          font-size="10"
-          x="32"
-          v-bind:y="`${BASE_Y_SIZE * zoomY * (index + 1) - 4}`"
-          v-if="y.pitch === 'C'"
-          class="sequencer-keys-item-pitchname"
-        >
-          {{ y.name }}
-        </text>
-      </g>
-    </svg>
+    <!-- 鍵盤 -->
+    <sequencer-keys />
+    <!-- シーケンサ -->
     <div class="sequencer-body">
       <!-- グリッド -->
       <!-- NOTE: 現状小節+オクターブごとの罫線なし -->
       <svg
-        v-bind:height="`${BASE_Y_SIZE * zoomY * 128}`"
-        v-bind:width="`${gridX.length * BASE_X_SIZE * zoomX}`"
+        v-bind:height="`${sizeY * zoomY * 128}`"
+        v-bind:width="`${gridX.length * sizeX * zoomX}`"
         xmlns="http://www.w3.org/2000/svg"
         class="sequencer-grids"
+        @click="(e) => addNote(e)"
       >
         <defs>
           <pattern
             id="sequencer-grid-16"
-            v-bind:width="`${BASE_X_SIZE * zoomX}px`"
-            v-bind:height="`${12 * BASE_Y_SIZE * zoomY}px`"
+            v-bind:width="`${sizeX * zoomX}px`"
+            v-bind:height="`${12 * sizeY * zoomY}px`"
             patternUnits="userSpaceOnUse"
           >
             <rect
               v-for="(y, index) in gridY"
               :key="index"
               x="0"
-              v-bind:y="`${BASE_Y_SIZE * zoomY * index}`"
-              v-bind:width="`${BASE_X_SIZE * zoomX}`"
-              v-bind:height="`${BASE_Y_SIZE * zoomY}`"
+              v-bind:y="`${sizeY * zoomY * index}`"
+              v-bind:width="`${sizeX * zoomX}`"
+              v-bind:height="`${sizeY * zoomY}`"
               v-bind:class="`sequencer-grids-col sequencer-grids-col-${y.color}`"
             />
           </pattern>
           <!-- NOTE: 4/4 1小節でグリッド見た目確認目的 -->
           <pattern
             id="sequencer-grid-measure"
-            v-bind:width="`${BASE_X_SIZE * 16 * zoomX}`"
-            v-bind:height="`${12 * BASE_Y_SIZE * zoomY}`"
+            v-bind:width="`${sizeX * 16 * zoomX}`"
+            v-bind:height="`${12 * sizeY * zoomY}`"
             patternUnits="userSpaceOnUse"
           >
             <rect width="100%" height="100%" fill="url(#sequencer-grid-16)" />
@@ -90,63 +56,14 @@
           height="100%"
           id="sequencer-grid"
           fill="url(#sequencer-grid-measure)"
-          @mousedown="(e) => addNote(e)"
         />
       </svg>
-      <!-- NOTE: ノートと歌詞入力あわせコンポーネント分割予定 -->
-      <div
+      <sequencer-note
         v-for="(note, index) in notes"
         :key="index"
-        class="sequencer-note"
-        v-bind:style="{
-          transform: `translate3d(${(note.position / 4) * zoomX}px,${
-            (127 - note.midi) * BASE_Y_SIZE * zoomY
-          }px,0)`,
-        }"
-      >
-        <!-- NOTE: q-inputへの変更 / 表示面+速度面から、クリック時のみinputを表示に変更予定 -->
-        <input
-          type="text"
-          :value="note.lyric"
-          @input="(e) => setLyric(index, e)"
-          class="sequencer-note-lyric"
-        />
-        <svg
-          v-bind:height="`${BASE_Y_SIZE * zoomY}`"
-          v-bind:width="`${(note.duration / 4) * zoomX}`"
-          xmlns="http://www.w3.org/2000/svg"
-          class="sequencer-note-bar"
-        >
-          <g @dblclick="removeNote(index)">
-            <rect
-              y="0"
-              x="0"
-              height="100%"
-              width="100%"
-              stroke-width="1"
-              class="sequencer-note-bar-body"
-            />
-            <rect
-              y="0"
-              x="-4"
-              height="100%"
-              width="8"
-              fill-opacity="0"
-              draggable
-              class="sequencer-note-bar-draghandle"
-            />
-            <rect
-              y="0"
-              v-bind:x="`${(note.duration / 4) * zoomX - 4}`"
-              height="100%"
-              width="8"
-              fill-opacity="0"
-              class="sequencer-note-bar-draghandle"
-              draggable
-            />
-          </g>
-        </svg>
-      </div>
+        v-bind:note="note"
+        v-bind:index="index"
+      />
     </div>
     <!-- NOTE: スクロールバー+ズームレンジ仮 -->
     <input
@@ -186,14 +103,22 @@
 <script lang="ts">
 import { defineComponent, computed, onMounted } from "vue";
 import { useStore } from "@/store";
+import SequencerKeys from "@/components/Sing/SequencerKeys.vue";
+import SequencerNote from "@/components/Sing/SequencerNote.vue";
 import {
   midiKeys,
   getPitchFromMidi,
   getDoremiFromMidi,
+  BASE_GRID_SIZE_X as sizeX,
+  BASE_GRID_SIZE_Y as sizeY,
 } from "@/helpers/singHelper";
 
 export default defineComponent({
   name: "SingScoreSequencer",
+  components: {
+    SequencerKeys,
+    SequencerNote,
+  },
   setup() {
     const store = useStore();
     const gridY = midiKeys;
@@ -220,22 +145,19 @@ export default defineComponent({
     const zoomY = computed(() => store.state.sequencerZoomY);
     const scrollX = computed(() => store.state.sequencerScrollX);
     const scrollY = computed(() => store.state.sequencerScrollY);
-    const BASE_Y_SIZE = 30;
-    const BASE_X_SIZE = 30;
     onMounted(() => {
       const el = document.querySelector("#score-sequencer");
       // C4あたりにスクロールする
       if (el) {
-        el.scrollTop = scrollY.value * (BASE_Y_SIZE * zoomY.value);
+        el.scrollTop = scrollY.value * (sizeY * zoomY.value);
       }
     });
     const addNote = (event: MouseEvent) => {
       const resolution = store.state.score?.resolution;
       const gridXSize = resolution ? resolution / 4 : 120;
       const position =
-        gridXSize * Math.floor(event.offsetX / (BASE_X_SIZE * zoomX.value));
-      const midi =
-        127 - Math.floor(event.offsetY / (BASE_Y_SIZE * zoomY.value));
+        gridXSize * Math.floor(event.offsetX / (sizeX * zoomX.value));
+      const midi = 127 - Math.floor(event.offsetY / (sizeX * zoomY.value));
       if (0 > midi) {
         return;
       }
@@ -250,26 +172,6 @@ export default defineComponent({
           lyric,
         },
       });
-    };
-    const removeNote = (index: number) => {
-      store.dispatch("REMOVE_NOTE", { index });
-    };
-    // NOTE: ノートのバーと歌詞入力でコンポーネント分割予定
-    const setLyric = (index: number, event: Event) => {
-      if (!(event.target instanceof HTMLInputElement)) {
-        return;
-      }
-      if (event.target.value && store.state.score) {
-        const currentNote = store.state.score?.notes[index];
-        const lyric = event.target.value;
-        store.dispatch("CHANGE_NOTE", {
-          index,
-          note: {
-            ...currentNote,
-            lyric,
-          },
-        });
-      }
     };
     const setZoomX = (event: Event) => {
       if (!(event.target instanceof HTMLInputElement)) {
@@ -294,14 +196,12 @@ export default defineComponent({
       notes,
       zoomX,
       zoomY,
+      sizeX,
+      sizeY,
       getPitchFromMidi,
       addNote,
-      removeNote,
-      setLyric,
       setZoomX,
       setZoomY,
-      BASE_X_SIZE,
-      BASE_Y_SIZE,
     };
   },
 });
@@ -319,35 +219,6 @@ export default defineComponent({
   padding-bottom: 164px;
   position: relative;
   width: 100%;
-}
-
-.sequencer-keys {
-  backface-visibility: hidden;
-  background: #fff;
-  display: block;
-  position: sticky;
-  left: 0;
-  z-index: 100;
-}
-
-.sequencer-keys-item-separator-octave {
-  stroke: #bbb;
-}
-
-.sequencer-keys-item-separator-f {
-  stroke: #ddd;
-}
-
-.sequencer-keys-item-white {
-  fill: #fff;
-}
-
-.sequencer-keys-item-black {
-  fill: #555;
-}
-
-.sequencer-keys-item-pitchname {
-  fill: #555;
 }
 
 .sequencer-body {
@@ -382,47 +253,5 @@ export default defineComponent({
 .sequencer-grid-separator-line {
   backface-visibility: hidden;
   stroke: #b0b0b0;
-}
-
-.sequencer-note {
-  backface-visibility: hidden;
-  position: absolute;
-  top: 0;
-  left: 0;
-  transform-origin: 0, 0;
-}
-
-.sequencer-note-lyric {
-  background: white;
-  border: 0;
-  border-bottom: 1px solid colors.$primary-light;
-  color: colors.$display;
-  font-size: 12px;
-  font-weight: bold;
-  font-feature-settings: "palt" 1;
-  letter-spacing: 0.05em;
-  outline: none;
-  padding: 0;
-  position: absolute;
-  bottom: calc(100% - 1px);
-  left: 0;
-  width: 24px;
-}
-
-.sequencer-note-bar {
-  display: block;
-  position: relative;
-}
-.sequencer-note-bar-body {
-  fill: colors.$primary;
-  stroke: #fff;
-  stroke-opacity: 0.5;
-  position: relative;
-  top: 0;
-  left: 0;
-}
-
-.sequencer-note-bar-draghandle {
-  cursor: ew-resize;
 }
 </style>
